Rename `services` to `layers`
