+++ crates/fj-core/src/lib.rs
@@ -83,10 +83,10 @@
 
 pub mod algorithms;
 pub mod geometry;
+pub mod layers;
 pub mod objects;
 pub mod operations;
 pub mod queries;
-pub mod services;
 pub mod storage;
 pub mod validate;
